# docs/swagger.yaml
definitions:
  handler.request:
    properties:
      count:
        type: integer
      name:
        type: string
      price:
        type: number
      type:
        type: string
    required:
    - count
    - name
    - price
    - type
    type: object
  web.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
    type: object
info:
  contact:
    name: API Support
    url: http://developers.mercadolibre.com.ar/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://developers.mercadolibre.com.ar/es_ar/terminos_y_condiciones
  title: MELI Bootcamp API
  version: "1.0"
paths:
  /products:
    get:
      consumes:
      - application/json
      description: get products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: store products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product to store
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.Response'
      tags:
      - Products
swagger: "2.0"